moved query next packet sequence
