+++ x/stakeibc/client/cli/query.go
@@ -238,3 +238,32 @@ func CmdShowEpochTracker() *cobra.Command {
 
 	return cmd
 }
+
+func CmdNextPacketSequence() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "next-packet-sequence [channel-id] [port-id]",
+		Short: "returns the next packet sequence on a channel",
+		Args:  cobra.ExactArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx := client.GetClientContextFromCmd(cmd)
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			channelId := args[0]
+			portId := args[1]
+
+			params := &types.QueryGetNextPacketSequenceRequest{ChannelId: channelId, PortId: portId}
+
+			res, err := queryClient.NextPacketSequence(context.Background(), params)
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}
